requirements.txt
langchain
langchain-community
langchain-openai
langchain-postgres
langgraph
psycopg2-binary
python-dotenv
fastapi
uvicorn
pydantic
pandas
sqlalchemy
langchain-core